dj_b_ecomm_project/requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
certifi==2025.10.5
chardet==3.0.4
charset-normalizer==3.4.4
deep-translator==1.11.4
Django==5.2.7
django-modeltranslation==0.19.17
django-rosetta==0.10.2
django-widget-tweaks==1.5.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pillow==12.0.0
polib==1.2.0
requests==2.32.5
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
